Split native shell state regression tests
