+++ src/app_core/native_shell/composition/state.rs
@@ -412,524 +412,5 @@ impl NativeShellState {
 }
 
 #[cfg(test)]
-mod opt_272_tests {
-    use super::*;
-    use crate::app_core::native_shell::runtime_contract::{
-        FolderPaneIdModel, SourceRowModel, folder_row_model,
-    };
-    use crate::gui::types::Vector2;
-
-    fn browser_model_with_rows(total: usize, focused_visible_row: usize) -> AppModel {
-        let mut model = AppModel::default();
-        for visible_row in 0..total {
-            model.browser.rows.push(BrowserRowModel::new(
-                visible_row,
-                format!("row_{visible_row:04}"),
-                1,
-                false,
-                visible_row == focused_visible_row,
-            ));
-        }
-        model.browser.visible_count = model.browser.rows.len();
-        model.browser.selected_visible_row = Some(focused_visible_row);
-        model.browser.anchor_visible_row = Some(focused_visible_row.saturating_sub(2));
-        model.browser.autoscroll = true;
-        model
-    }
-
-    fn folder_model_with_rows(total_rows: usize, focused_row: usize) -> AppModel {
-        let mut model = AppModel::default();
-        model.sources.rows.push(SourceRowModel::new(
-            String::from("source"),
-            String::from("detail"),
-            true,
-            false,
-        ));
-        model.sources.upper_folder_pane.active = true;
-        model.sources.upper_folder_pane.has_item = true;
-        model.sources.upper_folder_pane.focused_tree_row = Some(focused_row);
-        model.sources.active_folder_pane = FolderPaneIdModel::Upper;
-        for row_index in 0..total_rows {
-            model
-                .sources
-                .upper_folder_pane
-                .tree_rows
-                .push(folder_row_model(
-                    format!("folder_{row_index:03}"),
-                    String::new(),
-                    row_index % 3,
-                    false,
-                    row_index == focused_row,
-                    row_index == 0,
-                    row_index + 1 < total_rows,
-                    true,
-                ));
-        }
-        model
-    }
-
-    /// Build a populated single-sidebar fixture for source/folder geometry checks.
-    fn populated_single_sidebar_model() -> AppModel {
-        let mut model = folder_model_with_rows(48, 4);
-        model.sources.rows.clear();
-        for index in 0..12 {
-            model.sources.rows.push(SourceRowModel::new(
-                format!("source_{index:02}"),
-                format!("detail_{index:02}"),
-                index == 4,
-                false,
-            ));
-        }
-        model
-    }
-
-    #[test]
-    /// The sidebar reserves one source list and one folder browser at all densities.
-    fn sidebar_sections_render_one_source_and_folder_browser_across_viewports() {
-        let sizes = [
-            Vector2::new(820.0, 520.0),
-            Vector2::new(1280.0, 720.0),
-            Vector2::new(2300.0, 1080.0),
-        ];
-        let mut state = NativeShellState::new();
-        let model = populated_single_sidebar_model();
-        for viewport in sizes {
-            let layout = ShellLayout::build(viewport);
-            let style = style_for_layout(&layout);
-            let sections = sidebar_sections(&layout, &style, &model);
-            let rendered_sources = state.rendered_source_row_rects(&layout, &model);
-            let expected_source_rows = rendered_source_rows(&style, &model);
-            assert!(sections.upper.bounds.height() > sections.lower.bounds.height());
-            assert!(sections.lower.bounds.height() <= 0.01);
-            assert_eq!(rendered_sources.len(), expected_source_rows);
-        }
-    }
-
-    /// Compact sidebar workspace keeps sources, tags, and filters ordered.
-    #[test]
-    fn compact_sidebar_workspace_anchors_tags_and_filters_below_sources() {
-        let model = populated_single_sidebar_model();
-        for viewport in [Vector2::new(820.0, 420.0), Vector2::new(1280.0, 720.0)] {
-            let layout = ShellLayout::build(viewport);
-            let style = style_for_layout(&layout);
-            let workspace = sidebar_workspace_sections(&layout, &style);
-            let sections = sidebar_sections(&layout, &style, &model);
-
-            assert!(layout.sidebar_rows.contains(workspace.sources.center()));
-            assert!(layout.sidebar_rows.contains(workspace.tags.center()));
-            assert!(layout.sidebar_rows.contains(workspace.filters.center()));
-            assert!(workspace.sources.max.y <= workspace.tags.min.y);
-            assert!(workspace.tags.max.y <= workspace.filters.min.y);
-            assert!(workspace.sources.contains(sections.upper.bounds.center()));
-        }
-    }
-
-    /// Left-sidebar rating chips route through the source hit-test path.
-    #[test]
-    fn left_sidebar_rating_chip_routes_browser_filter_action() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let model = populated_single_sidebar_model();
-        let mut state = NativeShellState::new();
-        let chip = state
-            .sidebar_rating_filter_chip_rect(&layout, &model, 3)
-            .expect("left-sidebar rating chip should exist");
-
-        assert_eq!(
-            state.source_action_at_point(&layout, &model, chip.center()),
-            Some(
-                crate::app_core::native_shell::runtime_contract::UiAction::ToggleBrowserRatingFilter {
-                    level: 3,
-                    invert: false,
-                }
-            )
-        );
-    }
-
-    /// Left-sidebar metadata filter rows open dropdowns that route sidebar filter actions.
-    #[test]
-    fn left_sidebar_metadata_filter_dropdowns_route_browser_filter_actions() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let model = populated_single_sidebar_model();
-        let mut state = NativeShellState::new();
-
-        let expected = [
-            (
-                0,
-                0,
-                crate::app_core::app_api::state::BrowserSidebarFilterOption::Format(
-                    crate::app_core::app_api::state::BrowserFormatFacet::Wav,
-                ),
-            ),
-            (
-                1,
-                0,
-                crate::app_core::app_api::state::BrowserSidebarFilterOption::BitDepth(
-                    crate::app_core::app_api::state::BrowserBitDepthFacet::Unavailable,
-                ),
-            ),
-            (
-                2,
-                3,
-                crate::app_core::app_api::state::BrowserSidebarFilterOption::Channels(
-                    crate::app_core::app_api::state::BrowserChannelFacet::Unavailable,
-                ),
-            ),
-            (
-                3,
-                0,
-                crate::app_core::app_api::state::BrowserSidebarFilterOption::Bpm(
-                    crate::app_core::app_api::state::BrowserBpmFacet::Unknown,
-                ),
-            ),
-            (
-                4,
-                0,
-                crate::app_core::app_api::state::BrowserSidebarFilterOption::Key(
-                    crate::app_core::app_api::state::BrowserKeyFacet::Unknown,
-                ),
-            ),
-        ];
-
-        for (row_index, option_index, option) in expected {
-            let row = state
-                .sidebar_filter_row_rect(&layout, &model, row_index)
-                .expect("left-sidebar filter row should exist");
-            assert_eq!(
-                state.source_action_at_point(&layout, &model, row.center()),
-                Some(crate::app_core::native_shell::runtime_contract::UiAction::FocusBrowserPanel)
-            );
-            assert!(state.sidebar_filter_dropdown_visible());
-            let option_rect = state
-                .sidebar_filter_dropdown_option_rect(&layout, &model, option_index)
-                .expect("left-sidebar filter dropdown option should exist");
-            assert_eq!(
-                state.sidebar_filter_dropdown_action_at_point(
-                    &layout,
-                    &model,
-                    option_rect.center()
-                ),
-                Some(
-                    crate::app_core::native_shell::runtime_contract::UiAction::ToggleBrowserSidebarFilter {
-                        option,
-                        additive: true,
-                    }
-                )
-            );
-        }
-    }
-
-    /// Left-sidebar filter dropdowns expose clear actions for active facets.
-    #[test]
-    fn left_sidebar_filter_dropdown_clear_routes_browser_filter_action() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let mut model = populated_single_sidebar_model();
-        model.sidebar_filters.toggle(
-            crate::app_core::app_api::state::BrowserSidebarFilterOption::Bpm(
-                crate::app_core::app_api::state::BrowserBpmFacet::Mid,
-            ),
-            true,
-        );
-        let mut state = NativeShellState::new();
-        let row = state
-            .sidebar_filter_row_rect(&layout, &model, 3)
-            .expect("left-sidebar BPM filter row should exist");
-
-        assert_eq!(
-            state.source_action_at_point(&layout, &model, row.center()),
-            Some(crate::app_core::native_shell::runtime_contract::UiAction::FocusBrowserPanel)
-        );
-        let clear_rect = state
-            .sidebar_filter_dropdown_option_rect(&layout, &model, 4)
-            .expect("left-sidebar filter dropdown clear option should exist");
-
-        assert_eq!(
-            state.sidebar_filter_dropdown_action_at_point(&layout, &model, clear_rect.center()),
-            Some(
-                crate::app_core::native_shell::runtime_contract::UiAction::ClearBrowserSidebarFilter {
-                    facet: crate::app_core::app_api::state::BrowserSidebarFilterFacet::Bpm,
-                }
-            )
-        );
-    }
-
-    #[test]
-    /// The single visible folder browser keeps its scrollbar thumb hit target active.
-    fn single_folder_browser_scrollbar_thumb_is_hittable() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let style = style_for_layout(&layout);
-        let model = folder_model_with_rows(240, 72);
-        let mut state = NativeShellState::new();
-        let rows = state
-            .cached_tree_rows(&layout, &style, &model, FolderPaneIdModel::Upper)
-            .to_vec();
-        let sections = sidebar_sections(&layout, &style, &model);
-        let scrollbar = folder_scrollbar_layout(
-            sections.tree_rows(FolderPaneIdModel::Upper),
-            &rows,
-            model.sources.upper_folder_pane.tree_rows.len(),
-            style.sizing,
-        )
-        .expect("overflowing single folder browser should render a scrollbar");
-        let point = scrollbar.thumb.center();
-
-        let (slot, offset) = state
-            .folder_scrollbar_thumb_offset_at_point(&layout, &model, point)
-            .expect("single folder scrollbar thumb should be hittable");
-
-        assert_eq!(slot, FolderPaneIdModel::Upper);
-        assert!((offset - (scrollbar.thumb.height() * 0.5)).abs() <= 0.001);
-    }
-
-    #[test]
-    fn browser_rows_use_generic_list_window_hit_testing_and_scrollbar_primitives() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let style = style_for_layout(&layout);
-        let model = browser_model_with_rows(240, 118);
-        let mut state = NativeShellState::new();
-
-        let rows = state.cached_browser_rows(&layout, &style, &model).to_vec();
-        let list_rect = browser_rows_list_rect(layout.browser_rows, style.sizing, &model);
-        let expected_len = browser_rows_capacity(list_rect, style.sizing);
-
-        assert_eq!(rows.len(), expected_len);
-        assert!(rows.iter().any(|row| row.visible_row == 118));
-
-        let target = rows[3].rect.center();
-        assert_eq!(
-            row_index_for_visible_rows(&rows, target, list_rect),
-            Some(3)
-        );
-
-        let scrollbar =
-            browser_scrollbar_layout(list_rect, &rows, model.browser.visible_count, style.sizing)
-                .expect("overflowing browser rows should expose a scrollbar");
-        assert!(scrollbar.track.contains(scrollbar.thumb.center()));
-        assert_eq!(
-            browser_scrollbar_view_start_for_pointer(
-                scrollbar,
-                rows.len(),
-                model.browser.visible_count,
-                scrollbar.track.max.y,
-                scrollbar.thumb.height(),
-            ),
-            Some(model.browser.visible_count - rows.len())
-        );
-    }
-
-    #[test]
-    fn source_folder_rows_use_generic_list_window_and_scrollbar_primitives() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let style = style_for_layout(&layout);
-        let model = folder_model_with_rows(160, 112);
-        let mut state = NativeShellState::new();
-
-        let rows = state
-            .cached_tree_rows(&layout, &style, &model, FolderPaneIdModel::Upper)
-            .to_vec();
-        let sections = sidebar_sections(&layout, &style, &model);
-        let tree_rect = sections.tree_rows(FolderPaneIdModel::Upper);
-        let expected_len = tree_rows_capacity(tree_rect, style.sizing);
-
-        assert_eq!(rows.len(), expected_len);
-        assert!(rows.iter().any(|row| row.row_index == 112));
-
-        let scrollbar = folder_scrollbar_layout(
-            tree_rect,
-            &rows,
-            model.sources.upper_folder_pane.tree_rows.len(),
-            style.sizing,
-        )
-        .expect("overflowing source folders should expose a scrollbar");
-        assert!(scrollbar.track.contains(scrollbar.thumb.center()));
-        assert_eq!(
-            folder_scrollbar_view_start_for_pointer(
-                scrollbar,
-                rows.len(),
-                model.sources.upper_folder_pane.tree_rows.len(),
-                scrollbar.track.max.y,
-                scrollbar.thumb.height(),
-            ),
-            Some(model.sources.upper_folder_pane.tree_rows.len() - rows.len())
-        );
-    }
-
-    #[test]
-    fn waveform_scrollbar_thumb_tracks_zoomed_view_position() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let scrollbar = waveform_scrollbar_layout(layout.waveform_scrollbar_lane, 250_000, 500_000)
-            .expect("zoomed waveform view should render a scrollbar");
-
-        assert!(scrollbar.track.min.y >= layout.waveform_scrollbar_lane.min.y);
-        assert!(scrollbar.track.max.y <= layout.waveform_scrollbar_lane.max.y);
-        assert!(scrollbar.track.min.y >= layout.waveform_plot.max.y);
-        assert!(scrollbar.thumb.min.x > scrollbar.track.min.x);
-        assert!(scrollbar.thumb.max.x < scrollbar.track.max.x);
-        assert!(scrollbar.track.height() <= 3.0);
-    }
-
-    #[test]
-    fn waveform_scrollbar_hides_when_view_is_fully_zoomed_out() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-
-        assert!(waveform_scrollbar_layout(layout.waveform_scrollbar_lane, 0, 1_000_000).is_none());
-    }
-
-    #[test]
-    fn waveform_toolbar_renders_svg_backed_icon_images() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let mut model = AppModel::default();
-        model.transport_running = false;
-        let mut state = NativeShellState::new();
-        let play_rect = state
-            .waveform_toolbar_button_rect(&layout, &model, "Play")
-            .expect("play waveform toolbar button should be present");
-        let channel_rect = state
-            .waveform_toolbar_button_rect(&layout, &model, "Channel")
-            .expect("channel waveform toolbar button should be present");
-
-        let frame = state.build_frame(&layout, &model);
-
-        for button_rect in [play_rect, channel_rect] {
-            let image = frame
-                .primitives
-                .iter()
-                .find_map(|primitive| match primitive {
-                    Primitive::Image(image)
-                        if image.rect.min.x >= button_rect.min.x
-                            && image.rect.min.y >= button_rect.min.y
-                            && image.rect.max.x <= button_rect.max.x
-                            && image.rect.max.y <= button_rect.max.y =>
-                    {
-                        Some(image)
-                    }
-                    _ => None,
-                })
-                .expect("toolbar button should render an SVG-backed image primitive");
-            assert!(
-                image.image.pixels.chunks_exact(4).any(|rgba| rgba[3] > 0),
-                "toolbar SVG image should contain visible pixels"
-            );
-        }
-    }
-
-    #[test]
-    fn options_panel_omits_inner_title_drag_surface() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let style = style_for_layout(&layout);
-        let mut state = NativeShellState::new();
-        let model = AppModel {
-            options_panel: crate::app_core::native_shell::runtime_contract::OptionsPanelModel {
-                visible: true,
-                ..crate::app_core::native_shell::runtime_contract::OptionsPanelModel::default()
-            },
-            ..AppModel::default()
-        };
-        let initial = options_panel_layout(&layout, &style, &model)
-            .expect("visible options panel should resolve layout");
-        let grab = initial.title_rect.center();
-
-        assert!(initial.title.is_empty());
-        assert_eq!(initial.title_rect.height(), 0.0);
-        assert!(!state.begin_options_panel_drag(&layout, &model, grab));
-    }
-
-    #[test]
-    fn options_panel_drag_clamps_window_inside_shell_bounds() {
-        let layout = ShellLayout::build(Vector2::new(900.0, 520.0));
-        let style = style_for_layout(&layout);
-        let model = AppModel {
-            options_panel: crate::app_core::native_shell::runtime_contract::OptionsPanelModel {
-                visible: true,
-                ..crate::app_core::native_shell::runtime_contract::OptionsPanelModel::default()
-            },
-            ..AppModel::default()
-        };
-
-        let panel = options_panel_layout_for_origin(
-            &layout,
-            &style,
-            &model,
-            Some(Point::new(-10_000.0, 10_000.0)),
-        )
-        .expect("visible options panel should resolve layout");
-
-        assert!(panel.panel_rect.min.x >= layout.root.rect.min.x);
-        assert!(panel.panel_rect.min.y >= layout.top_bar.max.y);
-        assert!(panel.panel_rect.max.x <= layout.root.rect.max.x);
-        assert!(panel.panel_rect.max.y <= layout.status_bar.min.y);
-    }
-
-    #[test]
-    fn options_panel_picker_mode_expands_inline_dropdown_actions() {
-        let layout = ShellLayout::build(Vector2::new(1280.0, 720.0));
-        let style = style_for_layout(&layout);
-        let state = NativeShellState::new();
-        let model = AppModel {
-            options_panel: crate::app_core::native_shell::runtime_contract::OptionsPanelModel {
-                visible: true,
-                ..crate::app_core::native_shell::runtime_contract::OptionsPanelModel::default()
-            },
-            paired_device: crate::app_core::native_shell::runtime_contract::PairedDevicePanelModel {
-                active_picker: Some(
-                    crate::app_core::native_shell::runtime_contract::PairedPickerTargetModel::PrimaryNumber,
-                ),
-                primary_number: crate::app_core::native_shell::runtime_contract::SummaryFieldModel {
-                    label: String::from("Sample Rate"),
-                    value_label: String::from("48 kHz"),
-                },
-                primary_number_options: vec![
-                    crate::app_core::native_shell::runtime_contract::PairedPickerOptionModel {
-                        label: String::from("Device default"),
-                        selected: false,
-                        value: crate::app_core::native_shell::runtime_contract::PairedPickerValueModel::PrimaryNumber(None),
-                    },
-                    crate::app_core::native_shell::runtime_contract::PairedPickerOptionModel {
-                        label: String::from("48 kHz"),
-                        selected: true,
-                        value: crate::app_core::native_shell::runtime_contract::PairedPickerValueModel::PrimaryNumber(Some(
-                            48_000,
-                        )),
-                    },
-                ],
-                ..crate::app_core::native_shell::runtime_contract::PairedDevicePanelModel::default()
-            },
-            ..AppModel::default()
-        };
-
-        let panel = options_panel_layout(&layout, &style, &model)
-            .expect("visible picker panel should resolve layout");
-        assert!(panel.title.is_empty());
-        let dropdown_row = panel
-            .buttons
-            .iter()
-            .position(|button| button.action == UiAction::OpenPrimaryNumberPicker)
-            .expect("active picker row should remain in the overview");
-        assert!(panel.buttons[dropdown_row].active);
-        assert!(panel.buttons[dropdown_row].text.starts_with("Sample Rate"));
-        assert_eq!(
-            panel.buttons[dropdown_row + 1].action,
-            UiAction::SetPrimaryNumber { value: None }
-        );
-        assert_eq!(
-            panel.buttons[dropdown_row + 2].action,
-            UiAction::SetPrimaryNumber {
-                value: Some(48_000),
-            }
-        );
-        assert!(panel.buttons[dropdown_row + 2].active);
-
-        let dropdown_point = panel.buttons[dropdown_row].rect.center();
-        assert_eq!(
-            state.options_panel_action_at_point(&layout, &model, dropdown_point),
-            Some(UiAction::OpenPrimaryNumberPicker)
-        );
-
-        let sample_rate_point = panel.buttons[dropdown_row + 2].rect.center();
-        assert_eq!(
-            state.options_panel_action_at_point(&layout, &model, sample_rate_point),
-            Some(UiAction::SetPrimaryNumber {
-                value: Some(48_000),
-            })
-        );
-    }
-}
+#[path = "state/tests/opt_272.rs"]
+mod opt_272_tests;
